+++ aidial_sdk/_pydantic/_compat.py
