+++ aidial_sdk/chat_completion/request.py
@@ -5,11 +5,11 @@ import pydantic
 from pydantic import PositiveInt, StrictBool, StrictInt, StrictStr
 from typing_extensions import Annotated, assert_never
 
-from aidial_sdk._pydantic import Field
+from aidial_sdk._pydantic import PYDANTIC_V2, Field
+from aidial_sdk._pydantic._compat import model_validator
 from aidial_sdk.chat_completion.enums import Status
 from aidial_sdk.deployment.from_request_mixin import FromRequestDeploymentMixin
 from aidial_sdk.exceptions import InvalidRequestError
-from aidial_sdk.pydantic._compat import PYDANTIC_V2, model_validator
 from aidial_sdk.utils.pydantic import ExtraForbidModel
