+++ aidial_sdk/deployment/truncate_prompt.py
@@ -1,8 +1,8 @@
 from typing import List, Literal, Union
 
+from aidial_sdk._pydantic._compat import BaseModel
 from aidial_sdk.chat_completion.request import ChatCompletionRequest
 from aidial_sdk.deployment.from_request_mixin import FromRequestDeploymentMixin
-from aidial_sdk.pydantic._compat import BaseModel
 
 
 class TruncatePromptRequest(FromRequestDeploymentMixin):
