+++ aidial_sdk/utils/log_config.py
@@ -1,6 +1,6 @@
 import os
 
-from aidial_sdk.pydantic._compat import BaseModel
+from aidial_sdk._pydantic._compat import BaseModel
 
 DIAL_SDK_LOG = os.environ.get("DIAL_SDK_LOG", "WARNING").upper()
