+++ tests/test_pydantic.py
@@ -2,8 +2,8 @@ import pytest
 from pydantic import BaseModel as NativeBaseModel
 
 from aidial_sdk._pydantic import INSTALLED_PYDANTIC_V2, USE_PYDANTIC_V2
-from aidial_sdk.pydantic._compat import BaseModel as DialBaseModel
-from tests.utils._pydantic import model_dump
+from aidial_sdk._pydantic._compat import BaseModel as DialBaseModel
+from tests.utils.pydantic import model_dump
 
 pytestmark = pytest.mark.skipif(
     not USE_PYDANTIC_V2 and INSTALLED_PYDANTIC_V2,
