+++ tests/test_tokenize.py
@@ -3,7 +3,7 @@ from typing import List
 import pytest
 
 from aidial_sdk import DIALApp
-from aidial_sdk.pydantic._compat import PYDANTIC_V2
+from aidial_sdk._pydantic import PYDANTIC_V2
 from tests.applications.echo import EchoApplication
 from tests.applications.noop import NoopApplication
 from tests.utils.endpoint_test import TestCase, run_endpoint_test
